Plot data

+++ PostureVR/Assets/Scripts/DataCollection.cs
@@ -3,6 +3,14 @@ using System.Collections.Generic;
 using UnityEngine;
 using System.IO;
 
+public enum InterventionType
+{
+    Base,
+    Icon,
+    Dot,
+    Environment
+}
+
 public class DataCollection : MonoBehaviour
 {
     public PoorPostureDetection poorPostureDetection;
@@ -13,6 +21,8 @@ public class DataCollection : MonoBehaviour
 
     public int timesPerSecond;
 
+    public InterventionType interventionType;
+
     [System.Serializable]
 
     public class Player
@@ -21,6 +31,7 @@ public class DataCollection : MonoBehaviour
         public float angle;
         public int postureState;
         public int interventionTriggered; 
+        public InterventionType intervention;
     }
 
     [SerializeField]
@@ -42,14 +53,19 @@ public class DataCollection : MonoBehaviour
         if (startCollectingData)
         {
             timer += Time.fixedDeltaTime;
-            if (timer > 1f / timesPerSecond)
+            if (timer >= 1f / timesPerSecond)
             {
                 timer = 0f;
                 Player playerData = new Player();
                 playerData.height = Camera.main.transform.localPosition.y;
                 playerData.angle = Camera.main.transform.eulerAngles.x;
+                if (playerData.angle > 90f)
+                {
+                    playerData.angle = playerData.angle - 360f;
+                }
                 playerData.postureState = poorPostureDetection.m_isPoorPosture ? 1 : 0;
                 playerData.interventionTriggered = poorPostureDetection.interventionTriggered ? 1 : 0;
+                playerData.intervention = interventionType;
                 player.Add(playerData);
             }
         }
@@ -59,15 +75,19 @@ public class DataCollection : MonoBehaviour
     {
         if (player.Count > 0)
         {
-            TextWriter tw = new StreamWriter(filename, false);
-            tw.WriteLine("Height, Angle, State, Intervention");
-            tw.Close();
-
+            TextWriter tw;
+            if (!new FileInfo(filename).Exists)
+            {
+                tw = new StreamWriter(filename, false);
+                tw.WriteLine("Height, Angle, State, IsTriggered, Intervention");
+                tw.Close();
+            }
+            
             tw = new StreamWriter(filename, true);
 
             for (int i = 0; i < player.Count; ++i)
             {
-                tw.WriteLine(player[i].height + "," + player[i].angle + "," + player[i].postureState + "," + player[i].interventionTriggered);
+                tw.WriteLine(player[i].height + "," + player[i].angle + "," + player[i].postureState + "," + player[i].interventionTriggered + "," + player[i].intervention);
             }
             tw.Close();
         }
@@ -79,6 +99,7 @@ public class DataCollection : MonoBehaviour
         if (endCollectingData)
         {
             WriteCSV();
+            endCollectingData = false;
         }
     }
 }
